gcp-flask/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
uWSGI==2.0.18
Werkzeug==0.15.0
numpy==1.20.1
pandas==1.2.2
praat-parselmouth==0.4.0
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.1
six==1.15.0
